Replacing shorts with more meaningful types.

+++ app/src/main/java/com/dpanayotov/wallpaper/Bar.java
@@ -4,10 +4,10 @@ package com.dpanayotov.wallpaper;
  * Created by Dean Panayotov Local on 7.9.2015
  */
 public class Bar {
-    public short y;
+    public float y;
     public int color;
 
-    public Bar(short y, int color) {
+    public Bar(float y, int color) {
         this.y = y;
         this.color = color;
     }
